refactor: tidy up the extensions a bit
